+++ kolibri/core/assets/src/core-app/constructor.js
@@ -34,6 +34,8 @@ const publicMethods = [
   'off',
   'registerLanguageAssets',
   'registerLanguageAssetsUrl',
+  'registerContentRenderer',
+  'retrieveContentRenderer',
 ];
 
 /**