+++ kolibri/plugins/document_pdf_render/assets/src/module.js
@@ -6,8 +6,8 @@ class DocumentPDFModule extends ContentRendererModule {
   get rendererComponent() {
     return PDFComponent;
   }
-  get contentType() {
-    return 'document/pdf';
+  get contentTypes() {
+    return require('./content_types.json');
   }
 }