+++ kolibri/plugins/video_mp4_render/assets/src/module.js
@@ -6,8 +6,8 @@ class VideoMP4Module extends ContentRendererModule {
   get rendererComponent() {
     return VideoComponent;
   }
-  get contentType() {
-    return 'video/mp4';
+  get contentTypes() {
+    return require('./content_types.json');
   }
 }