+++ CHANGELOG.rst
@@ -39,3 +39,4 @@ Changed
 
  - Webpack build process now compatible with plugins housed outside the kolibri directory.
  - ES2015 transpilation now uses Bublé instead of Babel.
+ - Content Renderers now use explicit API different to standard asynchronous loading.